@@ -7,7 +7,7 @@ namespace Microsoft.eShopWeb.ApplicationCore.Interfaces
     {
         Task<int> GetBasketItemCountAsync(string userName);
         Task TransferBasketAsync(string anonymousId, string userName);
-        Task AddItemToBasket(int basketId, int catalogItemId, decimal price, int quantity);
+        Task AddItemToBasket(int basketId, int catalogItemId, decimal price, int quantity = 1);
         Task SetQuantities(int basketId, Dictionary<string, int> quantities);
         Task DeleteBasketAsync(int basketId);
     }
