@@ -4,6 +4,6 @@ namespace Microsoft.eShopWeb.Web.Controllers.Api
 {
     [Route("api/[controller]/[action]")]
     [ApiController]
-    public class BaseApiController : Controller
+    public class BaseApiController : ControllerBase
     { }
 }
