@@ -35,7 +35,7 @@
                             <section class="esh-basket-item esh-basket-item--middle col-xs-2">$ @item.UnitPrice.ToString("N2")</section>
                             <section class="esh-basket-item esh-basket-item--middle col-xs-2">
                                 <input type="hidden" name="@("Items[" + i + "].Key")" value="@item.Id" />
-                                <input type="number" class="esh-basket-input" min="1" name="@("Items[" + i + "].Value")" value="@item.Quantity" />
+                                <input type="number" class="esh-basket-input" min="0" name="@("Items[" + i + "].Value")" value="@item.Quantity" />
                             </section>
                             <section class="esh-basket-item esh-basket-item--middle esh-basket-item--mark col-xs-2">$ @Math.Round(item.Quantity * item.UnitPrice, 2).ToString("N2")</section>
                         </div>
@@ -87,8 +87,12 @@
     }
     else
     {
-        <div class="esh-catalog-items row">
-            Basket is empty.   <a asp-page="/Index">Continue Shopping..</a>
-        </div>
+        <h3 class="esh-catalog-items row">
+            Basket is empty.
+        </h3>
+        
+        <section class="esh-basket-item">
+            <a asp-page="/Index" class="btn esh-basket-checkout text-white">[ Continue Shopping..]</a>
+        </section>
     }
 </div>
