@@ -5,13 +5,14 @@ using Microsoft.eShopWeb.ApplicationCore.Entities.BasketAggregate;
 using Moq;
 using System;
 using Xunit;
+using System.Threading.Tasks;
 
 namespace Microsoft.eShopWeb.UnitTests.ApplicationCore.Services.BasketServiceTests
 {
     public class SetQuantities
     {
-        private int _invalidId = -1;
-        private Mock<IAsyncRepository<Basket>> _mockBasketRepo;
+        private readonly int _invalidId = -1;
+        private readonly Mock<IAsyncRepository<Basket>> _mockBasketRepo;
 
         public SetQuantities()
         {
@@ -19,7 +20,7 @@ namespace Microsoft.eShopWeb.UnitTests.ApplicationCore.Services.BasketServiceTes
         }
 
         [Fact]
-        public async void ThrowsGivenInvalidBasketId()
+        public async Task ThrowsGivenInvalidBasketId()
         {
             var basketService = new BasketService(_mockBasketRepo.Object, null);
 
@@ -28,7 +29,7 @@ namespace Microsoft.eShopWeb.UnitTests.ApplicationCore.Services.BasketServiceTes
         }
 
         [Fact]
-        public async void ThrowsGivenNullQuantities()
+        public async Task ThrowsGivenNullQuantities()
         {
             var basketService = new BasketService(null, null);
 
